requirements.txt
discord~=2.3.2
python-dotenv
SQLAlchemy
aiosqlite
greenlet